VOTE/requirements.txt
django==3.2.16
gunicorn==20.1.0
whitenoise==5.3.0
psycopg2-binary==2.8.6
django-allauth==0.51.0
django-environ==0.8.1
django-sslserver==0.22
django-debug-toolbar==3.7.0
requests==2.28.1
requests-oauthlib==1.3.1
urllib3==1.26.12
django-session-timeout==0.1.0
djangorestframework==3.13.0
django_tables2==2.7.3
